chore: update dependencies and manage them with the bom plugin
Updating the versions of plugin dependencies manually can be exhausting,
looking for the correct set of versions that work with each other and do
not cause upper-bounds dependency error reports.

The Jenkins plugin bill of materials simplifies that by shifting the
evaluation of versions from each plugin author to a centralized process
maintained by members of the Jenkins community.