fix: use imageID instead of imageDigest and update to 1.22.1 (#131)
Locally built images don't have a image digest, since they are not
coming from a registry. This commit changes the usage of imageDigest
to imageID so we can also create reports for locally built images.

This also updates the version of the CLI to 1.22.1 by default to be
up-to-date with the latest changes.